middleware-api/python_agent/requirements.txt
# ========================================
# HealthLink AI Agent - Core Dependencies
# ========================================

# LangGraph and LangChain Core
langgraph>=1.0.0
langchain-core>=1.0.0
langchain-google-genai>=3.0.0

# Google Gemini LLM Dependencies (versions auto-resolved by pip)
google-generativeai
google-ai-generativelanguage

# Environment and Configuration
python-dotenv>=1.0.0

# Pydantic for data validation
pydantic>=2.0.0

# ========================================
# Optional: For future Supabase/Web3 integration
# ========================================
# supabase>=2.0.0
# web3>=6.11.0
# aiohttp>=3.9.0
